requirements.txt
requests==2.27.1
pandas==1.5.3
psycopg2-binary==2.9.6